Move training and sorting of initial population into GP constructor so that we only sort once per gen at the end.
Add mutation target parameter which specifies whether to mutate the children, the non-parents or both. Also add mutation rate which is the probability of mutating any given mutation target.

Change the sorting method to using inplace list.sort. Change _best_individuals accumulator to be a LifoQueue so that it is thread-safe.

Add a __hash__method to individuals which is based on the __str__ method of that individual. This allows us to remove duplicates from the population with list(set(pop)).

Change initialisation methods to work on sets of individuals to ensure that no duplicates are generated. Convert the pop to a list before returning.